pyproject.toml
[project]
name = "aquestalk-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.124.0",
    "python-dotenv>=1.2.1",
    "uvicorn[standard]>=0.38.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
]
